requirements.txt
Flask
python-dotenv
flask_sqlalchemy
flask-mail 
Flask-Login 
cx-Oracle